Refactor organization for map components
