+++ frontend/src/components/map/basic/Arch.tsx
