+++ frontend/src/components/map/basic/Box.tsx
