+++ frontend/src/components/map/basic/MapComponent.tsx
@@ -1,8 +1,10 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import React from 'react';
 import { Vector3 } from '@react-three/fiber';
-import { Direction } from '../../types/map';
-import { HexCode } from '../../types/color';
-import { getRotation } from '../../utils/direction';
+import { Direction } from '../../../types/map';
+import { HexCode } from '../../../types/color';
+import { getRotation } from '../../../utils/direction';
 
 export interface Color {
   color: HexCode;
