+++ frontend/src/components/map/basic/Stairs.tsx
