+++ frontend/src/components/map/index.tsx
@@ -1,10 +1,10 @@
-export { default as Arch } from './Arch';
-export { default as Box } from './Box';
-export { default as Character } from './Character';
-export { default as Columns } from './Columns';
-export { default as Ladder } from './Ladder';
-export { default as Model } from './Model';
-export { default as Stairs } from './Stairs';
-export { default as Disc } from './Disc';
-export { default as NextDisc } from './NextDisc';
-export { default as Island } from './Island';
+export { default as Arch } from './basic/Arch';
+export { default as Box } from './basic/Box';
+export { default as Character } from './composite/Character';
+export { default as Columns } from './basic/Columns';
+export { default as Ladder } from './basic/Ladder';
+export { default as Model } from './basic/Model';
+export { default as Stairs } from './basic/Stairs';
+export { default as Disc } from './basic/Disc';
+export { default as NextDisc } from './composite/NextDisc';
+export { default as Island } from './composite/Island';
