+++ frontend/src/components/map/mapTemplates/SpaceMap.tsx
@@ -3,10 +3,10 @@ import React, { useState, useRef, Suspense } from 'react';
 import SpaceMapStructure from './SpaceMapStructure';
 import { Canvas } from '@react-three/fiber';
 import { MapControls, Stars } from '@react-three/drei';
-import { Character } from '../../components/map';
+import { Character } from '..';
 import { EffectComposer, Bloom } from '@react-three/postprocessing';
-import { Axis, Direction } from '../../types/map';
-import { DirectionPosition } from '../../types/map';
+import { Axis, Direction } from '../../../types/map';
+import { DirectionPosition } from '../../../types/map';
 import { getDirectionVector, nextDirectionCW } from 'utils/direction';
 import { translate } from 'utils/map';
 import { ChallengeMapData } from 'types/challenges';
