--- frontend/src/pages/maps/DemoMapTheme.ts
@@ -1,15 +1,0 @@
-import { createTheme } from '@mui/material';
-
-// Doesn't actually work yet.
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: '#ff6ad5',
-    },
-    secondary: {
-      main: '#8690ff',
-    },
-  },
-});
-
-export default theme;
