+++ frontend/src/pages/maps/MapsPage.tsx
@@ -3,11 +3,11 @@ import { batch, useDispatch, useSelector } from 'react-redux';
 import { loadFriendsOnSameChallenges } from 'store/challenges/operations';
 import { getChallengeMaps } from 'store/challenges/selectors';
 // import { useParams } from 'react-router-dom';
-import MapCarousel from './MapCarousel';
+import MapCarousel from '../../components/map/MapCarousel';
 import { RootState } from 'store';
 import { default as Appbar } from '../../components/common/Appbar';
 
-const MapsPage = () => {
+const MapsPage: React.FC = () => {
   // const { challengeId } = useParams<{ challengeId: string }>();
   const dispatch = useDispatch();
   useEffect(() => {
