+++ frontend/src/pages/maps/index.tsx
@@ -1,3 +1,1 @@
-export { default as DemoMap } from './DemoMap';
-export { default as SpaceMap } from './SpaceMap';
-export { default as MapsPage } from './MapsPage';
+export { default } from './MapsPage';
