+++ frontend/src/routing/routes.ts
@@ -12,10 +12,11 @@ import LoginPage from '../pages/login';
 import OnboardingPage from 'pages/onboarding/OnboardingPage';
 import ProfilePage from '../pages/profile';
 import RegistrationPage from '../pages/registration';
-import { DemoMap, MapsPage } from '../pages/maps';
+import MapsPage from '../pages/maps';
 import { RouteEntry } from '../types/routes';
 import AddFriendsPage from '../pages/friends/AddFriendsPage';
 import NotificationsPage from '../pages/notifications/NotificationsPage';
+import DemoMap from 'components/map/mapTemplates/DemoMap';
 
 export const HOME_ROUTE = '/';
 export const LOGIN_ROUTE = '/login';
